Rename OpenAI key modal component
