+++ src/components/talentforge/Settings.tsx
@@ -15,7 +15,7 @@ import {
   Typography,
 } from "@mui/material";
 import ErrorBoundary from "@/components/talentforge/ErrorBoundary";
-import OpenAIKeyModal from "@/components/talentforge/OpenAiKeyModal";
+import OpenAIKeyModal from "@/components/talentforge/OpenAIKeyModal";
 import { loadDemoData, clearDemoData } from "@/utils/talentforge/demoData";
 import { useTalentForgeData } from "@/contexts/TalentForgeDataContext";
 import { exportSnapshot, importSnapshot } from "@/utils/talentforge/snapshot";
